+++ renderer/src/web/price-check/filters/create-stat-filters.ts
@@ -53,7 +53,10 @@ export function createExactStatFilters (
     item.category !== ItemCategory.Sentinel
   )) {
     keepByType.push(ModifierType.Explicit)
-  } else if (item.rarity === ItemRarity.Rare && item.category === ItemCategory.Idol) {
+  } else if (
+    item.rarity === ItemRarity.Rare &&
+    (item.category === ItemCategory.Idol || item.category === ItemCategory.Chart)
+  ) {
     keepByType.push(ModifierType.Explicit)
   }
 
@@ -63,7 +66,7 @@ export function createExactStatFilters (
 
   const ctx: FiltersCreationContext = {
     item,
-    searchInRange: (item.category !== ItemCategory.Map)
+    searchInRange: (item.category !== ItemCategory.Map && item.category !== ItemCategory.Chart)
       ? Math.min(2, opts.searchStatRange)
       : opts.searchStatRange,
     filters: [],
@@ -87,7 +90,7 @@ export function createExactStatFilters (
     applyMirroredTabletRules(ctx.filters)
     return ctx.filters
   }
-  if (item.category === ItemCategory.Map) {
+  if (item.category === ItemCategory.Map || item.category === ItemCategory.Chart) {
     for (const filter of ctx.filters) {
       if (filter.tag !== FilterTag.Property && filter.tag !== FilterTag.Pseudo) {
         filter.disabled = false
@@ -204,16 +207,11 @@ export function calculatedStatToFilter (
       tag: (type === ModifierType.Enchant)
         ? FilterTag.Enchant
         : FilterTag.Variant,
-      oils: decodeOils(calc),
       sources: sources,
       option: {
         value: sources[0].contributes!.value
       },
       disabled: false
-    }
-
-    if (filter.oils) {
-      filter.disabled = true
     }
   }
 
@@ -232,6 +230,13 @@ export function calculatedStatToFilter (
     sources: sources,
     roll: undefined,
     disabled: true
+  }
+
+  if (calc.stat.better === StatBetter.NotComparable) {
+    if (type !== ModifierType.Enchant) {
+      filter.tag = FilterTag.Variant
+    }
+    filter.disabled = false
   }
 
   if (type === ModifierType.Implicit) {
@@ -269,6 +274,10 @@ export function calculatedStatToFilter (
       // filter.tag = FilterTag.Unveiled
     } else if (sources.some(s => CLIENT_STRINGS.INCURSION_MODS.includes(s.modifier.info.name!))) {
       filter.tag = FilterTag.Incursion
+    } else if (sources.some(s => CLIENT_STRINGS.ESSENCE_MODS.includes(s.modifier.info.name!))) {
+      filter.tag = FilterTag.Essence
+    } else if (sources.some(s => CLIENT_STRINGS.INFAMOUS_MODS.includes(s.modifier.info.name!))) {
+      filter.tag = FilterTag.Infamous
     }
   }